+++ Utils.h
@@ -17,14 +17,14 @@ void        rot(         mat_L& M,   const ZZX& f );
 void        rot_T(    mat_zz_p& M, const zz_pX& f );
 void        rot_vect( mat_zz_p& R, const vec_zz_pX& v );
 
-void        CoeffsX(vec_ZZ& coeffs_x, const vec_ZZX& x, const unsigned int& l);
-void        CoeffsInv(vec_zz_pX& x, const vec_ZZ& c, const unsigned int& l);
-void        CoeffsInvX(vec_ZZX& x, const vec_ZZ& c, const unsigned int& l);
-void        CoeffsHat(vec_ZZ& coeffs_x, const vec_ZZX& x, const unsigned int& l);
-void        CoeffsInvHat(vec_zz_pX& x, const vec_zz_p& c, const unsigned int& l);
-void        CoeffsInvHatX(vec_ZZX& x, const vec_ZZ& c, const unsigned int& l);
+void        CoeffsX(vec_ZZ& coeffs_x, const vec_ZZX& x, const unsigned long& l);
+void        CoeffsInv(vec_zz_pX& x, const vec_ZZ& c, const unsigned long& l);
+void        CoeffsInvX(vec_ZZX& x, const vec_ZZ& c, const unsigned long& l);
+void        CoeffsHat(vec_ZZ& coeffs_x, const vec_ZZX& x, const unsigned long& l);
+void        CoeffsInvHat(vec_zz_pX& x, const vec_zz_p& c, const unsigned long& l);
+void        CoeffsInvHatX(vec_ZZX& x, const vec_ZZ& c, const unsigned long& l);
 
-void        sigma_map(vec_zz_pX& N, const vec_zz_pX& M, const unsigned int& d);
+void        sigma_map(vec_zz_pX& N, const vec_zz_pX& M, const unsigned long& d);
 
 zz_pX       poly_mult(     const vec_zz_pX& f, const vec_zz_pX& g );
 zz_pX       poly_mult_hat( const vec_zz_pX& f, const vec_zz_pX& g );
@@ -32,7 +32,7 @@ zz_pX       poly_mult_hat( const vec_zz_pX& f, const vec_zz_pX& g );
 zz_pX       Compute_f(     const mat_zz_p& B_f, const ZZ& x );
 
 ZZ          Norm2(         const vec_ZZ&  v );
-ZZ          Norm2X(        const vec_ZZX& v, const unsigned int& d );
+ZZ          Norm2X(        const vec_ZZX& v, const long& d );
 double      Norm2D(        const vec_D&   v );
 
 double      InnerProdD(    const vec_D& a, const vec_D& b );
